monitoring/service-controller/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
docker==6.1.3
psutil==5.9.6
aiofiles==23.2.1